+++ src/main/java/znu/visum/components/movies/usecases/create/domain/CreateMovieService.java
@@ -3,5 +3,5 @@ package znu.visum.components.movies.usecases.create.domain;
 import znu.visum.components.movies.domain.models.Movie;
 
 public interface CreateMovieService {
-  Movie saveWithNameFromPeople(Movie movie);
+  Movie saveMovie(Movie movie);
 }