+++ src/test/java/znu/visum/components/externals/models/TmdbCreditsPeopleUnitTest.java
@@ -14,6 +14,8 @@ class TmdbCreditsPeopleUnitTest {
   @Nested
   class ToDomainExternalDirector {
 
+    private static final String BASE_POSTER_URL = "https://fake-url.io";
+
     @Test
     @DisplayName("When the people is not a director")
     void itShouldThrow() {
@@ -21,8 +23,9 @@ class TmdbCreditsPeopleUnitTest {
       crewPeople.setId(1);
       crewPeople.setName("Jacques Dupont");
       crewPeople.setJob("Original Music Composer");
+      crewPeople.setProfilePath("/poster1.jpg");
 
-      assertThatThrownBy(crewPeople::toDomainExternalDirector)
+      assertThatThrownBy(() -> crewPeople.toDomainWithBasePosterUrl(BASE_POSTER_URL))
           .isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -33,75 +36,98 @@ class TmdbCreditsPeopleUnitTest {
       crewPeople.setId(1);
       crewPeople.setName("Jacques Dupont");
       crewPeople.setJob("Director");
+      crewPeople.setProfilePath("/poster1.jpg");
 
-      assertThat(crewPeople.toDomainExternalDirector())
+      assertThat(crewPeople.toDomainWithBasePosterUrl(BASE_POSTER_URL))
           .usingRecursiveComparison()
-          .isEqualTo(new ExternalDirector(1, "Jacques", "Dupont"));
+          .isEqualTo(
+              new ExternalDirector(1, "Jacques", "Dupont", BASE_POSTER_URL + "/poster1.jpg"));
     }
 
     @Test
     @DisplayName("When the director has a name made of a single word")
-    void itShouldReturnAnDirectorWithAnEmptyName() {
+    void itShouldReturnDirectorWithAnEmptyName() {
       TmdbCrewPeople crewPeople = new TmdbCrewPeople();
       crewPeople.setId(1);
       crewPeople.setName("Jacques");
       crewPeople.setJob("Director");
+      crewPeople.setProfilePath("/poster1.jpg");
 
-      assertThat(crewPeople.toDomainExternalDirector())
+      assertThat(crewPeople.toDomainWithBasePosterUrl(BASE_POSTER_URL))
           .usingRecursiveComparison()
-          .isEqualTo(new ExternalDirector(1, "Jacques", ""));
+          .isEqualTo(new ExternalDirector(1, "Jacques", "", BASE_POSTER_URL + "/poster1.jpg"));
     }
 
     @Test
     @DisplayName("When the director has an empty name")
-    void itShouldReturnAnDirectorWithAnEmptyNameAndForename() {
+    void itShouldReturnDirectorWithAnEmptyNameAndForename() {
       TmdbCrewPeople crewPeople = new TmdbCrewPeople();
       crewPeople.setId(1);
       crewPeople.setName("");
       crewPeople.setJob("Director");
+      crewPeople.setProfilePath("/poster1.jpg");
 
-      assertThat(crewPeople.toDomainExternalDirector())
+      assertThat(crewPeople.toDomainWithBasePosterUrl(BASE_POSTER_URL))
           .usingRecursiveComparison()
-          .isEqualTo(new ExternalDirector(1, "", ""));
+          .isEqualTo(new ExternalDirector(1, "", "", BASE_POSTER_URL + "/poster1.jpg"));
     }
 
     @Test
     @DisplayName("When the director has a name composed of more than 2 words separated by spaces")
-    void itShouldReturnAnDirectorWithLeftPartAsForenameAndRightPartAsName() {
+    void itShouldReturnDirectorWithLeftPartAsForenameAndRightPartAsName() {
       TmdbCrewPeople crewPeople = new TmdbCrewPeople();
       crewPeople.setId(1);
       crewPeople.setName("Jacques Dupont Baguette Camembert");
       crewPeople.setJob("Director");
+      crewPeople.setProfilePath("/poster1.jpg");
 
-      assertThat(crewPeople.toDomainExternalDirector())
+      assertThat(crewPeople.toDomainWithBasePosterUrl(BASE_POSTER_URL))
           .usingRecursiveComparison()
-          .isEqualTo(new ExternalDirector(1, "Jacques", "Dupont Baguette Camembert"));
+          .isEqualTo(
+              new ExternalDirector(
+                  1, "Jacques", "Dupont Baguette Camembert", BASE_POSTER_URL + "/poster1.jpg"));
     }
 
     @Test
     @DisplayName("When the director has a name composed one word but ending with a space")
-    void itShouldReturnAnDirectorWithATrimForenameAndAnEmptyName() {
+    void itShouldReturnDirectorWithATrimForenameAndAnEmptyName() {
       TmdbCrewPeople crewPeople = new TmdbCrewPeople();
       crewPeople.setId(1);
       crewPeople.setName("Jacques ");
       crewPeople.setJob("Director");
+      crewPeople.setProfilePath("/poster1.jpg");
 
-      assertThat(crewPeople.toDomainExternalDirector())
+      assertThat(crewPeople.toDomainWithBasePosterUrl(BASE_POSTER_URL))
           .usingRecursiveComparison()
-          .isEqualTo(new ExternalDirector(1, "Jacques", ""));
+          .isEqualTo(new ExternalDirector(1, "Jacques", "", BASE_POSTER_URL + "/poster1.jpg"));
     }
 
     @Test
     @DisplayName("When the director has a name starting with a space")
-    void itShouldReturnAnDirectorWithATrimForename() {
+    void itShouldReturnDirectorWithATrimForename() {
       TmdbCrewPeople crewPeople = new TmdbCrewPeople();
       crewPeople.setId(1);
       crewPeople.setName(" Jacques");
       crewPeople.setJob("Director");
+      crewPeople.setProfilePath("/poster1.jpg");
 
-      assertThat(crewPeople.toDomainExternalDirector())
+      assertThat(crewPeople.toDomainWithBasePosterUrl(BASE_POSTER_URL))
           .usingRecursiveComparison()
-          .isEqualTo(new ExternalDirector(1, "Jacques", ""));
+          .isEqualTo(new ExternalDirector(1, "Jacques", "", BASE_POSTER_URL + "/poster1.jpg"));
+    }
+
+    @Test
+    @DisplayName("When the director no poster path")
+    void itShouldReturnDirectorWithNullPosterUrl() {
+      TmdbCrewPeople crewPeople = new TmdbCrewPeople();
+      crewPeople.setId(1);
+      crewPeople.setName(" Jacques");
+      crewPeople.setJob("Director");
+      crewPeople.setProfilePath(null);
+
+      assertThat(crewPeople.toDomainWithBasePosterUrl(BASE_POSTER_URL))
+          .usingRecursiveComparison()
+          .isEqualTo(new ExternalDirector(1, "Jacques", "", null));
     }
   }
 }