+++ src/test/resources/sql/insert_single_director.sql
@@ -1,2 +1,2 @@
-INSERT INTO director (id, name, forename)
-values (1, 'Lynch', 'David');
+INSERT INTO director (id, name, forename, tmdb_id, poster_url)
+values (1, 'Lynch', 'David', 1234, 'https://fakeurl.com');